Bugfix, jumps and new tests
Changed Instruction's opcode field to uint16_t to fix a bug with longer opcodes.
Added tests for the jumps and add/sub r/m64, r64

+++ include/instruction.h
@@ -9,7 +9,7 @@
 */
 
 struct Instruction{
-    uint8_t opcode;             // Unique internal representation of the instrution opcode
+    uint16_t opcode;             // Unique internal representation of the instrution opcode
     uint8_t operand_type;       // Operand types byte
     uint8_t op1;                // First operand (could be register ID, memory base reg, etc.)
     uint8_t op2;                // Second operand (same format as the first operand)
